Project/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
bcrypt==4.0.1
python-dotenv==1.0.0
Pillow==11.0.0
Werkzeug==2.3.7